update constructor with new syntax
